move MyMessageDialog into stddlib.h header
